项目完结

+++ 慕课网java学习代码/数据库基础/OA/src/main/java/com/xiaxinyu/OA/entity/Employee.java
@@ -1,17 +1,17 @@
 package com.xiaxinyu.OA.entity;
 
 public class Employee {
-    private Integer employeeId;
+    private Long employeeId;
     private String name;
     private Long departmentId;
     private String title;
     private Integer level;
 
-    public Integer getEmployeeId() {
+    public Long getEmployeeId() {
         return employeeId;
     }
 
-    public void setEmployeeId(Integer employeeId) {
+    public void setEmployeeId(Long employeeId) {
         this.employeeId = employeeId;
     }
